+++ src/main/java/entity/Portfolio.java
@@ -13,30 +13,10 @@ public class Portfolio {
 
     /** Constructor for Portfolio class.
      * Portfolio contains all the stocks a user is holding: tickers -> quantities(position).
-     * @param stocks a map of tickers and the number of shares the user holds.     *
+     * @param stocks a map of tickers and the stocks user hold
      */
-    // TODO: change the input to Map<String, Stock/UserStock> stocks
-    // 异常处理有待商榷
-    public Portfolio(Map<String, Integer> stocks) {
-        this.stocks = new HashMap<>();
-        StockMarket stockMarket = StockMarket.Instance();
-
-        stocks.forEach((ticker, quantity) -> {
-            // decompose Optional<Stock>
-            stockMarket.getStock(ticker).ifPresentOrElse(
-                stock -> {
-                    double price = stock.getPrice();
-                    if (price > 0) {
-                        this.stocks.put(ticker, new UserStock(stock, price, quantity));
-                    } else {
-                        throw new IllegalArgumentException("Invalid price for stock: " + ticker);
-                    }
-                },
-                () -> {
-                    throw new IllegalArgumentException("No stock found for ticker: " + ticker);
-                }
-            );
-        });
+    public Portfolio(Map<String, UserStock> stocks) {
+        this.stocks = new HashMap<>(stocks);
     }
 
     public double getTotalValue() {